+++ CHANGELOG.md
@@ -31,6 +31,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed Lara's Home entry not working.
 * Fixed exploding TR3 bosses.
 * Fixed original issue with deactivation of Dart Emitter.
+* Fixed original issue with weapon hotkeys available in binoculars or lasersight mode.
 * Fixed Lens Flare object not functioning properly.
 * Fixed lens flares not being occluded by static meshes and moveables.
 * Fixed spotlight shadows.
@@ -39,6 +40,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 ### New Features
 * Added fast savegame reloading.
 * Added ricochet sounds and make the effect more prominent.
+* Allow camera shake during flybys.
 * Allow to run the engine without title level.
 * Allow more than 1024 objects in a level.
 * Allow more than 1000 static mesh slots in a level.